Implement primary functionality: ispunct
Implements the primary functionality of the library: the `ispunct`
function, which takes a character (or, rather, a Python string of
length 1), and returns the all-important boolean.

The hardest (and potentially unnecessary) part of this was---based on
Julia's implementation of the same function---implementing the
`ismalformed` function, which does bitwise operations on the character
to check if it is an allowed Unicode codepoint, or if it is malformed.
I can't see a reason why this would come ever in using the present
library, but it was a fun exercise in bitwise operations and faux-
fixed-length integers in Python.