Improve error messages on bad input
There are a handful of `TypeError`s we can raise given a bad input.
The most obvious one is that the user gives the `ispunct` function a
non-string datum.  The others are inferred type errors, as Python does
not have a character literal type, so we must infer whether the input
can be treated as a character based on its length, or if it an invalid
string.